requirements.txt
fastapi
uvicorn
boto3
pandas
openpyxl  # Required for reading .xlsx files
ai4bharat-transliteration==1.1.3
alabaster==0.7.16
antlr4-python3-runtime==4.8
babel==2.14.0
bitarray==2.9.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cython==3.0.8
docutils==0.20.1
fairseq==0.12.2
filelock==3.13.1
flask==3.0.2
flask-cors==4.0.0
fsspec==2024.2.0
gevent==24.2.1
greenlet==3.0.3
hydra-core==1.0.7
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.1
indic-nlp-library==0.92
itsdangerous==2.1.2
jinja2==3.1.3
joblib==1.3.2
lxml==5.1.0
markupsafe==2.1.5
mock==5.1.0
morfessor==2.0.6
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.0.6
packaging==23.2
pandas==2.2.0
portalocker==2.8.2
progressbar2==4.3.2
protobuf==4.25.3
pyarrow==15.0.0
pycparser==2.21
pydload==1.0.9
pygments==2.17.2
python-dateutil==2.8.2
python-utils==3.8.2
pytz==2024.1
pyyaml==6.0.1
regex==2023.12.25
requests==2.31.0
sacrebleu==2.4.0
sacremoses==0.1.1
six==1.16.0
snowballstemmer==2.2.0
sphinx==7.2.6
sphinx-argparse==0.4.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sympy==1.12
tabulate==0.9.0
tensorboardx==2.6.2.2
torch==2.2.0
torchaudio==2.2.0
tqdm==4.66.2
transliterate==1.10.2
triton==2.2.0
typing-extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
urduhack==0.1.4
urllib3==2.2.1
werkzeug==3.0.1
zipp==3.17.0
zope-event==5.0
zope-interface==6.2
